Refactor: Representations into a seperate class to be included via composition
