+++ BoxContentSDK/BoxContentSDK/BOXContentSDKConstants.h
@@ -515,3 +515,33 @@ typedef NS_ENUM(NSUInteger, BOXAvatarType) {
     BOXAvatarTypeLarge
 };
 
+/**
+ Representations are digital assets created for files stored in Box. To request representations information for
+ a file, set the options for each type.
+ 
+ - BOXRepresentationRequestOriginal:                        Request original content url with file information request
+ Reference the reponse field 'authenticated_download_url' to download original content when file
+ has download permissions are owner, co-owner, editor, viewer uploader, viewer
+ - BOXRepresentationRequestAllRepresentations:              If permissions allow retrieve representations for thumbnails, preview representations for the given file
+ - BOXRepresentationRequestHighDefinitionVideo:             Request video representions if available for the given file
+ - BOXRepresentationRequestThumbnailRepresentation:         Request thumbnail representions if available for the given file
+ - BOXRepresentationRequestLargeThumbnailRepresentation:    Request large thumbnail representions are available for the given file
+ - BOXRepresentationRequestPDFRepresentation:               Request PDF representions if available for the given file
+ - BOXRepresentationRequestJPGRepresentation:               Request JPG representions if available for the given file
+ - BOXRepresentationRequestMP3Representation:               Request PNG representions if available for the given file
+ - BOXRepresentationRequestMP4Representation:               Request MP4 representions if available for the given file
+ - BOXRepresentationRequesteExtractedTextRepresentation:    Request extracted text if unformatted text is contained in a document (non-image)
+ */
+typedef NS_ENUM(NSUInteger, BOXRepresentationRequestOptions) {
+    BOXRepresentationRequestOriginal,
+    BOXRepresentationRequestAllRepresentations,
+    BOXRepresentationRequestHighDefinitionVideo,
+    BOXRepresentationRequestThumbnailRepresentation,
+    BOXRepresentationRequestLargeThumbnailRepresentation,
+    BOXRepresentationRequestPDFRepresentation,
+    BOXRepresentationRequestJPGRepresentation,
+    BOXRepresentationRequestMP3Representation,
+    BOXRepresentationRequestMP4Representation,
+    BOXRepresentationRequesteExtractedTextRepresentation
+} NS_ENUM_AVAILABLE_IOS(10_0);
+
